--- src/main/java/com/salesianostriana/dam/projectFOODAPP/categoria/Controller/CategoriaController.java
@@ -1,4 +1,0 @@
-package com.salesianostriana.dam.projectFOODAPP.categoria.Controller;
-
-public class CategoriaController {
-}
