--- src/main/java/com/salesianostriana/dam/projectFOODAPP/categoria/Dto/GetDtoCategoria.java
@@ -1,4 +1,0 @@
-package com.salesianostriana.dam.projectFOODAPP.categoria.Dto;
-
-public record GetDtoCategoria() {
-}
