+++ src/main/java/com/salesianostriana/dam/projectFOODAPP/categoria/controller/CategoriaController.java
@@ -1,0 +1,4 @@
+package com.salesianostriana.dam.projectFOODAPP.categoria.controller;
+
+public class CategoriaController {
+}
