+++ src/main/java/com/salesianostriana/dam/projectFOODAPP/categoria/dto/GetDtoCategoria.java
@@ -1,0 +1,4 @@
+package com.salesianostriana.dam.projectFOODAPP.categoria.dto;
+
+public record GetDtoCategoria() {
+}
