--- src/main/java/com/salesianostriana/dam/projectFOODAPP/pedido/Controller/PedidoController.java
@@ -1,4 +1,0 @@
-package com.salesianostriana.dam.projectFOODAPP.pedido.Controller;
-
-public class PedidoController {
-}
