--- src/main/java/com/salesianostriana/dam/projectFOODAPP/pedido/Dto/GetDtoPerido.java
@@ -1,4 +1,0 @@
-package com.salesianostriana.dam.projectFOODAPP.pedido.Dto;
-
-public record GetDtoPerido() {
-}
