+++ src/main/java/com/salesianostriana/dam/projectFOODAPP/pedido/controller/PedidoController.java
@@ -1,0 +1,4 @@
+package com.salesianostriana.dam.projectFOODAPP.pedido.controller;
+
+public class PedidoController {
+}
