+++ src/main/java/com/salesianostriana/dam/projectFOODAPP/pedido/dto/GetDtoPerido.java
@@ -1,0 +1,4 @@
+package com.salesianostriana.dam.projectFOODAPP.pedido.dto;
+
+public record GetDtoPerido() {
+}
