--- src/main/java/com/salesianostriana/dam/projectFOODAPP/producto/Controller/ProductoController.java
@@ -1,4 +1,0 @@
-package com.salesianostriana.dam.projectFOODAPP.producto.Controller;
-
-public class ProductoController {
-}
