--- src/main/java/com/salesianostriana/dam/projectFOODAPP/producto/Dto/GetDtoProducto.java
@@ -1,4 +1,0 @@
-package com.salesianostriana.dam.projectFOODAPP.producto.Dto;
-
-public record GetDtoProducto() {
-}
