+++ src/main/java/com/salesianostriana/dam/projectFOODAPP/producto/controller/ProductoController.java
@@ -1,0 +1,4 @@
+package com.salesianostriana.dam.projectFOODAPP.producto.controller;
+
+public class ProductoController {
+}
