+++ src/main/java/com/salesianostriana/dam/projectFOODAPP/producto/dto/GetDtoProducto.java
@@ -1,0 +1,4 @@
+package com.salesianostriana.dam.projectFOODAPP.producto.dto;
+
+public record GetDtoProducto() {
+}
