--- src/main/java/com/salesianostriana/dam/projectFOODAPP/usuario/Controller/UsuarioController.java
@@ -1,4 +1,0 @@
-package com.salesianostriana.dam.projectFOODAPP.usuario.Controller;
-
-public class UsuarioController {
-}
