--- src/main/java/com/salesianostriana/dam/projectFOODAPP/usuario/Dto/GetDtoCliente.java
@@ -1,4 +1,0 @@
-package com.salesianostriana.dam.projectFOODAPP.usuario.Dto;
-
-public record GetDtoCliente() {
-}
