+++ src/main/java/com/salesianostriana/dam/projectFOODAPP/usuario/controller/UsuarioController.java
@@ -1,0 +1,4 @@
+package com.salesianostriana.dam.projectFOODAPP.usuario.controller;
+
+public class UsuarioController {
+}
