+++ src/main/java/com/salesianostriana/dam/projectFOODAPP/usuario/dto/GetDtoCliente.java
@@ -1,0 +1,4 @@
+package com.salesianostriana.dam.projectFOODAPP.usuario.dto;
+
+public record GetDtoCliente() {
+}
